Split large handler.go file
handler.go became too big. Main changes:
- Hub-related functions have been moved to hub.go. In future we can decouple it from Cagent object and reuse in other places (like csender, or even frontman)
- SMART and VMs monitoring measurements collection moved to separate file similar to other measurements
- Functions in handler.go slightly reordered and access mode updated (public->private).
- collectMeasurements function implementation slightly simplified
- one function moved from config.go to handler.go as this is the only place where it's used

No new code or behaviour added.
