+++ handler.go
@@ -1,17 +1,11 @@
 package cagent
 
 import (
-	"bytes"
-	"compress/gzip"
 	"context"
-	"crypto/tls"
 	"encoding/json"
-	"io/ioutil"
 	"net/http"
-	"net/url"
 	"os"
 	"runtime"
-	"strings"
 	"time"
 
 	"github.com/pkg/errors"
@@ -23,183 +17,30 @@ import (
 	"github.com/cloudradar-monitoring/cagent/pkg/monitoring/networking"
 	"github.com/cloudradar-monitoring/cagent/pkg/monitoring/sensors"
 	"github.com/cloudradar-monitoring/cagent/pkg/monitoring/services"
-	"github.com/cloudradar-monitoring/cagent/pkg/monitoring/vmstat"
-	vmstatTypes "github.com/cloudradar-monitoring/cagent/pkg/monitoring/vmstat/types"
 )
 
-func (ca *Cagent) initHubClientOnce() {
-	ca.hubClientOnce.Do(func() {
-		transport := &http.Transport{
-			ResponseHeaderTimeout: 15 * time.Second,
-		}
-
-		rootCAs, err := common.CustomRootCertPool()
+func (ca *Cagent) Run(outputFile *os.File, interrupt chan struct{}) {
+	for {
+		err := ca.RunOnce(outputFile, ca.Config.OperationMode == OperationModeFull)
 		if err != nil {
-			if err != common.ErrorCustomRootCertPoolNotImplementedForOS {
-				log.Errorf("failed to add root certs: %s", err.Error())
-			}
-		} else if rootCAs != nil {
-			transport.TLSClientConfig = &tls.Config{
-				RootCAs: rootCAs,
-			}
+			log.Error(err)
 		}
 
-		if len(ca.Config.HubProxy) > 0 {
-			if !strings.HasPrefix(ca.Config.HubProxy, "http://") {
-				ca.Config.HubProxy = "http://" + ca.Config.HubProxy
-			}
-			proxyURL, err := url.Parse(ca.Config.HubProxy)
-			if err != nil {
-				log.WithFields(log.Fields{
-					"url": ca.Config.HubProxy,
-				}).Warningln("failed to parse hub_proxy URL")
-			} else {
-				if len(ca.Config.HubProxyUser) > 0 {
-					proxyURL.User = url.UserPassword(ca.Config.HubProxyUser, ca.Config.HubProxyPassword)
-				}
-				transport.Proxy = func(_ *http.Request) (*url.URL, error) {
-					return proxyURL, nil
-				}
-			}
+		select {
+		case <-interrupt:
+			return
+		case <-time.After(secToDuration(ca.Config.Interval)):
+			continue
 		}
-		ca.hubClient = &http.Client{
-			Timeout:   time.Duration(ca.Config.HubRequestTimeout) * time.Second,
-			Transport: transport,
-		}
-	})
+	}
 }
 
-// validateHubURL performs Hub URL validation, that reference field name as in source config.
-func (ca *Cagent) validateHubURL(fieldHubURL string) error {
-	if len(ca.Config.HubURL) == 0 {
-		return newEmptyFieldError(fieldHubURL)
-	} else if u, err := url.Parse(ca.Config.HubURL); err != nil {
-		err = errors.WithStack(err)
-		return newFieldError(fieldHubURL, err)
-	} else if u.Scheme != "http" && u.Scheme != "https" {
-		err := errors.Errorf("wrong scheme '%s', URL must start with http:// or https://", u.Scheme)
-		return newFieldError(fieldHubURL, err)
-	}
-	return nil
+func (ca *Cagent) RunOnce(outputFile *os.File, fullMode bool) error {
+	measurements := ca.collectMeasurements(fullMode)
+	return ca.reportMeasurements(measurements, outputFile)
 }
 
-// CheckHubCredentials performs credentials check for a Hub config, returning errors that reference
-// field names as in source config. Since config may be filled from file or UI, the field names can be different.
-// Consider also localization of UI, we want to decouple credential checking logic from their actual view in UI.
-//
-// Examples:
-// * for TOML: CheckHubCredentials(ctx, "hub_url", "hub_user", "hub_password")
-// * for WinUI: CheckHubCredentials(ctx, "URL", "User", "Password")
-func (ca *Cagent) CheckHubCredentials(ctx context.Context, fieldHubURL, fieldHubUser, fieldHubPassword string) error {
-	ca.initHubClientOnce()
-	err := ca.validateHubURL(fieldHubURL)
-	if err != nil {
-		return err
-	}
-
-	req, _ := http.NewRequest("HEAD", ca.Config.HubURL, nil)
-	req.Header.Add("User-Agent", ca.userAgent())
-	if len(ca.Config.HubUser) > 0 {
-		req.SetBasicAuth(ca.Config.HubUser, ca.Config.HubPassword)
-	}
-
-	ctx, cancelFn := context.WithTimeout(ctx, time.Minute)
-	req = req.WithContext(ctx)
-	resp, err := ca.hubClient.Do(req)
-	cancelFn()
-	if err = ca.checkClientError(resp, err, fieldHubUser, fieldHubPassword); err != nil {
-		return errors.WithStack(err)
-	}
-
-	return nil
-}
-
-func (ca *Cagent) checkClientError(resp *http.Response, err error, fieldHubUser, fieldHubPassword string) error {
-	if err != nil {
-		if errors.Cause(err) == context.DeadlineExceeded {
-			err = errors.New("connection timeout, please check your proxy or firewall settings")
-			return err
-		}
-		return err
-	}
-
-	var responseBody string
-	responseBodyBytes, readBodyErr := ioutil.ReadAll(resp.Body)
-	if readBodyErr == nil {
-		responseBody = string(responseBodyBytes)
-	}
-
-	_ = resp.Body.Close()
-	if resp.StatusCode == http.StatusUnauthorized {
-		if len(ca.Config.HubUser) == 0 {
-			return newEmptyFieldError(fieldHubUser)
-		} else if len(ca.Config.HubPassword) == 0 {
-			return newEmptyFieldError(fieldHubPassword)
-		}
-		return errors.Errorf("unable to authorize with provided Hub credentials (HTTP %d). %s", resp.StatusCode, responseBody)
-	} else if resp.StatusCode < http.StatusOK || resp.StatusCode >= http.StatusBadRequest {
-		return errors.Errorf("got unexpected response from server (HTTP %d). %s", resp.StatusCode, responseBody)
-	}
-	return nil
-}
-
-func newEmptyFieldError(name string) error {
-	err := errors.Errorf("unexpected empty field %s", name)
-	return errors.Wrap(err, "the field must be filled with details of your Cloudradar account")
-}
-
-func newFieldError(name string, err error) error {
-	return errors.Wrapf(err, "%s field verification failed", name)
-}
-
-func (ca *Cagent) PostResultToHub(ctx context.Context, result *Result) error {
-	ca.initHubClientOnce()
-	err := ca.validateHubURL("hub_url")
-	if err != nil {
-		return err
-	}
-
-	b, err := json.Marshal(result)
-	if err != nil {
-		err = errors.Wrap(err, "failed to serialize result")
-		return err
-	}
-
-	var req *http.Request
-	if ca.Config.HubGzip {
-		buf := new(bytes.Buffer)
-		gzipped := gzip.NewWriter(buf)
-		if _, err := gzipped.Write(b); err != nil {
-			err = errors.Wrap(err, "failed to write into gzipped buffer")
-			return err
-		}
-		if err := gzipped.Close(); err != nil {
-			err = errors.Wrap(err, "failed to finalize gzipped buffer")
-			return err
-		}
-		req, err = http.NewRequest("POST", ca.Config.HubURL, buf)
-		req.Header.Set("Content-Encoding", "gzip")
-	} else {
-		req, err = http.NewRequest("POST", ca.Config.HubURL, bytes.NewBuffer(b))
-	}
-	if err != nil {
-		return errors.WithStack(err)
-	}
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Add("User-Agent", ca.userAgent())
-	if len(ca.Config.HubUser) > 0 {
-		req.SetBasicAuth(ca.Config.HubUser, ca.Config.HubPassword)
-	}
-	req = req.WithContext(ctx)
-	resp, err := ca.hubClient.Do(req)
-	if err = ca.checkClientError(resp, err, "hub_user", "hub_password"); err != nil {
-		return errors.WithStack(err)
-	}
-
-	return nil
-}
-
-func (ca *Cagent) CollectMeasurements(full bool) (common.MeasurementsMap, error) {
+func (ca *Cagent) collectMeasurements(fullMode bool) common.MeasurementsMap {
 	var errCollector = common.ErrorCollector{}
 	var measurements = make(common.MeasurementsMap)
 
@@ -225,7 +66,7 @@ func (ca *Cagent) CollectMeasurements(full bool) (common.MeasurementsMap, error)
 		)
 	}
 
-	if full {
+	if fullMode {
 		info, err := ca.HostInfoResults()
 		errCollector.Add(err)
 		measurements = measurements.AddWithPrefix("system.", info)
@@ -299,22 +140,19 @@ func (ca *Cagent) CollectMeasurements(full bool) (common.MeasurementsMap, error)
 		}
 	}
 
+	measurements["operation_mode"] = ca.Config.OperationMode
+
 	if errCollector.HasErrors() {
+		measurements["message"] = errCollector.Combine()
 		measurements["cagent.success"] = 0
 	} else {
 		measurements["cagent.success"] = 1
 	}
 
-	measurements["operation_mode"] = ca.Config.OperationMode
-
-	if errCollector.HasErrors() {
-		return measurements, errCollector.Combine()
-	}
-
-	return measurements, nil
+	return measurements
 }
 
-func (ca *Cagent) ReportMeasurements(measurements common.MeasurementsMap, outputFile *os.File) error {
+func (ca *Cagent) reportMeasurements(measurements common.MeasurementsMap, outputFile *os.File) error {
 	result := &Result{
 		Timestamp:    time.Now().Unix(),
 		Measurements: measurements,
@@ -322,8 +160,7 @@ func (ca *Cagent) ReportMeasurements(measurements common.MeasurementsMap, output
 	if outputFile != nil {
 		err := json.NewEncoder(outputFile).Encode(result)
 		if err != nil {
-			err = errors.Wrap(err, "failed to JSON encode measurement result")
-			return err
+			return errors.Wrap(err, "failed to JSON encode measurement result")
 		}
 		return nil
 	}
@@ -341,33 +178,6 @@ func (ca *Cagent) ReportMeasurements(measurements common.MeasurementsMap, output
 	}
 
 	return err
-}
-
-func (ca *Cagent) RunOnce(outputFile *os.File, full bool) error {
-	measurements, err := ca.CollectMeasurements(full)
-	if err != nil {
-		// don't need to log or return it here – just add the message to report
-		// it is already logged (down into the CollectMeasurements)
-		measurements["message"] = err.Error()
-	}
-
-	return ca.ReportMeasurements(measurements, outputFile)
-}
-
-func (ca *Cagent) Run(outputFile *os.File, interrupt chan struct{}) {
-	for {
-		err := ca.RunOnce(outputFile, ca.Config.OperationMode == OperationModeFull)
-		if err != nil {
-			log.Error(err)
-		}
-
-		select {
-		case <-interrupt:
-			return
-		case <-time.After(secToDuration(ca.Config.Interval)):
-			continue
-		}
-	}
 }
 
 func (ca *Cagent) RunHeartbeat(interrupt chan struct{}) {
@@ -430,51 +240,6 @@ func (ca *Cagent) prettyPrintMeasurementsToFile(measurements common.Measurements
 	}
 }
 
-func (ca *Cagent) getVMStatMeasurements(f func(string, common.MeasurementsMap, error)) {
-	ca.vmstatLazyInit.Do(func() {
-		if err := vmstat.Init(); err != nil {
-			log.Error("vmstat: cannot instantiate virtual machines API: ", err.Error())
-			return
-		}
-
-		for _, name := range ca.Config.VirtualMachinesStat {
-			vm, err := vmstat.Acquire(name)
-			if err != nil {
-				if err != vmstatTypes.ErrNotAvailable {
-					log.Warnf("vmstat: Error while acquiring vm provider \"%s\": %s", name, err.Error())
-				}
-			} else {
-				ca.vmWatchers[name] = vm
-			}
-		}
-	})
-
-	for name, p := range ca.vmWatchers {
-		res, err := p.GetMeasurements()
-		f(name, res, err)
-	}
-}
-
-func (ca *Cagent) getSMARTMeasurements() common.MeasurementsMap {
-	// measurements fetched below should not affect cagent.success
-	if ca.smart != nil {
-		res, errs := ca.smart.Parse()
-
-		if len(errs) > 0 {
-			var errStr []string
-			for _, e := range errs {
-				errStr = append(errStr, e.Error())
-			}
-
-			if res == nil {
-				res = make(common.MeasurementsMap)
-			}
-
-			res["messages"] = strings.Join(errStr, "; ")
-		}
-
-		return res
-	}
-
-	return nil
+func secToDuration(seconds float64) time.Duration {
+	return time.Duration(int64(float64(time.Second) * seconds))
 }
